requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.9
django-phonenumber-field==6.0.0
django-redis==5.0.0
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
kombu==5.2.2
phonenumbers==8.12.38
Pillow==8.4.0
prompt-toolkit==3.0.23
psycopg2-binary==2.9.2
PyJWT==2.3.0
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
